Tab closing behavior has changed.
Once a tab is right clicked, the tab is made the currently selected tab.
Closing the currently selected tab now results in the selection of the
last remaining tab.

All new tabs now sport an "Open file" button.  Clicking this button
presents a file dialog.  When a file is opened through this dialog,
the file content occupies the tab and the "Open file" button disappears.
The newly opened file's filename becomes the label of the tab.